requirements.txt
pymupdf
pdfplumber
chromadb
sentence-transformers
rank-bm25
langchain
langchain-community
langchain-google-genai
google-generativeai
ragas
mlflow
fastapi
uvicorn
streamlit
python-dotenv